AP-1078 Refactor CCMS ref on means/merits reports
The CCMS case reference is not appearing on means and merits reports.
This was working but has stopped, due to a change in the CCMS submission
process. Previously the call to CCMS to get the case reference always
took place before the reports were generated but this is no longer the
case.

To fix this, move the ensure_case_ccms_reference_exists functionality
from the means/merits_report_controllers (where it wasn't being called)
into the means/merits_report_creator services, and ensure a CCMS case
reference is always present before creating the reports.

This includes a refactor of the report_creators, creating a
base_report_creator to provide common functionality.
