+++ app/services/reports/means_report_creator.rb
@@ -1,15 +1,5 @@
 module Reports
-  class MeansReportCreator
-    def self.call(legal_aid_application)
-      new(legal_aid_application).call
-    end
-
-    attr_reader :legal_aid_application
-
-    def initialize(legal_aid_application)
-      @legal_aid_application = legal_aid_application
-    end
-
+  class MeansReportCreator < BaseReportCreator
     def call
       return if legal_aid_application.means_report
 
@@ -25,11 +15,9 @@ module Reports
 
     private
 
-    def pdf_report
-      WickedPdf.new.pdf_from_string(html_report)
-    end
-
     def html_report
+      ensure_case_ccms_reference_exists
+
       Providers::MeansReportsController.default_url_options = {
         _recall: {
           legal_aid_application_id: legal_aid_application.id
