+++ spec/services/reports/means_report_creator_spec.rb
@@ -7,6 +7,7 @@ RSpec.describe Reports::MeansReportCreator do
 
   describe '.call' do
     it 'attaches means_report.pdf to the application' do
+      expect_any_instance_of(CCMS::ReferenceDataRequestor).to receive(:call)
       expect(Providers::MeansReportsController.renderer).to receive(:render).and_call_original
       subject
       legal_aid_application.reload
